Revert "delegando"
